feat: create element-types.ts
